[SWM-322] Refactor : remove unuse classses
