--- src/main/java/com/m9d/sroom/material/model/CourseQuizInfo.java
@@ -1,17 +1,0 @@
-package com.m9d.sroom.material.model;
-
-import lombok.Builder;
-import lombok.Data;
-
-@Data
-@Builder
-public class CourseQuizInfo {
-
-    private Long courseId;
-
-    private Long quizId;
-
-    private Long videoId;
-
-    private Long courseVideoId;
-}
