+++ src/main/java/com/m9d/sroom/repository/course/CourseJdbcRepositoryImpl.java
@@ -1,17 +1,12 @@
 package com.m9d.sroom.repository.course;
 
 import com.m9d.sroom.global.mapper.Course;
-import com.m9d.sroom.global.mapper.CourseVideo;
 import com.m9d.sroom.lecture.dto.response.CourseBrief;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.text.SimpleDateFormat;
 import java.util.List;
-import java.util.Optional;
-
-import static com.m9d.sroom.course.sql.CourseSqlQuery.GET_COURSE_LIST_QUERY;
 
 @Repository
 public class CourseJdbcRepositoryImpl implements CourseRepository {
