+++ src/test/java/com/m9d/sroom/lecture/controller/LectureResponseControllerTest.java
@@ -13,7 +13,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import static org.hamcrest.Matchers.*;
 
-public class LectureControllerTest extends ControllerTest {
+public class LectureResponseControllerTest extends ControllerTest {
 
 
     @Test
