+++ src/test/java/com/m9d/sroom/lecture/service/LectureResponseServiceTest.java
@@ -12,7 +12,7 @@ import java.util.List;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertTrue;
 
-public class LectureServiceTest extends ServiceTest {
+public class LectureResponseServiceTest extends ServiceTest {
 
     @Test
     void shouldReturnUnescapeHtml() {
